pyproject.toml
[tool.poetry]
name = "pouchdb-backup-script"
version = "0.1.0"
description = ""
authors = ["MrMaguila <gabrielgf6000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-storage = "^2.18.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
